Partially revert c6e5314fab3ee2e05590b69f578a4fb1ae1903e5
"An executable downloaded from an URL that works on machine X will also
work on machine Y" is not a good assumption, especially in multi-arch
environment.

This commit adds tags to the new tasks introduced in
c6e5314fab3ee2e05590b69f578a4fb1ae1903e5, so that they can be skipped.

The download task is also reverted to the previous behavior.